Moved BaseClient class to its own separate file
